Login page fixes
In order to fix an issue with timeouts the forms have been modified to only be presented by tomcat and not directly accessible to the user.  Now clicking the login button will send you to the login success message which is locked and requires login of at least user level priviledge.  Entering any other page with locks will direct you back after successfuly logging in.